Respect Bundler config
Currently Tara sets BUNDLE_IGNORE_CONFIG=1 when installing dependencies. This causes problems when the credentials for a gem repo is stored in the config and not in the Gemfile, see https://github.com/phusion/traveling-ruby/issues/53.

This changes so that by default this environment variable is not set, and a config option `:bundle_ignore_config` is added to set it.